pyproject.toml
[tool.poetry]
name = "pipeline-python-mongo-mysql"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pymongo = "^4.8.0"
pandas = "^2.2.2"
mysql-connector-python = "^9.0.0"
ipykernel = "^6.29.5"
requests = "^2.32.3"
setuptools = "^71.0.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
prospector = "^1.10.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
